add global notifications and remove deprecated props
