+++ packages/components/src/Notification/_docs/InlineNotification.mdx
@@ -41,8 +41,3 @@ import * as InlineNotificationExamples from "./InlineNotification.stories"
 
 <Description of={InlineNotificationExamples.ContentMultilineDemo} />
 <Canvas of={InlineNotificationExamples.ContentMultilineDemo} />
-
-### autohide
-
-<Description of={InlineNotificationExamples.AutohideDemo} />
-<Canvas of={InlineNotificationExamples.AutohideDemo} />
