--- packages/components/src/Notification/_mixins.scss
@@ -1,10 +1,0 @@
-// reset user agent styles for button element type
-@mixin button-reset {
-  appearance: none;
-  background: transparent;
-  border: none;
-  font: inherit;
-  margin: 0;
-  padding: 0;
-  transition: none;
-}
