--- packages/components/src/Notification/type.ts
@@ -1,6 +1,0 @@
-export type NotificationType =
-  | "positive"
-  | "informative"
-  | "cautionary"
-  | "negative"
-  | "security"
